requirements.txt
py==3.8.4
discord.py==1.3.4